refactor: New module game_restoration
